Move source code into sub-directory /src
