+++ src/ConsoleModifyBase.cpp
