+++ src/ConsoleModifyBase.h
